requirements.txt
Django==5.1.1
python-decouple==3.8
flake8==7.1.1
black==24.8.0
djangorestframework==3.15.2
Markdown==3.7
django-filter==24.3
django-unfold==0.39.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
django-prometheus==2.3.1
django-health-check==3.18.3
django-sphinxdoc==2.0.0
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
django-split-settings==1.3.2
pillow==10.4.0